Providing GPU enabled docker image
Providing CUDA and CUDNN driver within docker image simplify
compatibility issues.

This is especially true for onnxruntime 1.6 and 1.7 dependanding
respectively on CUDA 10 and 11.

Those images does not try to minimize the image size. I'll be happy to
add any simple optimization. But, as ML and DL tend to be heavy in size
I believe simplicity is probably more interesting than size.

Signed-off-by: haixuanTao <hai-xuan.tao@student.ecp.fr>